Add nonce
The `nonce` param is used to prevent replay attacks. The possible attack
scenario:

- An attacker has access to Dream app to be able to get a fresh `state`
  CSRF token.

- An attacker observes a redirect back to OIDC client (dream app) from
  OIDC provider which has authorization `code` in its response.

- An attacker replays the same request to the OIDC client (dream app)
  with acquired `code` and fresh `state` CSRF token and triggers OIDC
  client to make another call to get another access token / id tokens
  and thus be able to auth as legitimate user.

Now if we use have `nonce` (which is bound to legitimate's user session)
in token response then OIDC client (dream app) will not perform
authentication of the attacker (as they don't have access to the
legitimate's user session).

Note that this implementation only performs a check if `nonce` received
from the OIDC provider is bound to the current request session.

The stronger protection could be achieved by making `nonce` (or `state`)
to be **truly nonce** (persist it outside of user agent and invalidate
on first use). In this case even if attacker has access to browser
session the replay won't be possible.